requirements.txt
bidict==0.23.1
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pycparser==2.22
pyserial==3.5
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
werkzeug==3.0.4
wsproto==1.2.0
zipp==3.20.2
zope.event==5.0
zope.interface==7.0.3
